server/pyproject.toml
[tool.poetry.scripts]

[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = ["highlandscyclist <highlands.cyclist@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.8"
django-environ = "^0.4.5"
Django = "^3.1.7"
graphene-django = "^2.15.0"
django-graphql-jwt = "0.3.1"
django-cors-headers = "^3.7.0"
django-filter = "^2.4.0"
psycopg2-binary = "^2.8.6"
django-extensions = "^3.1.1"
PyJWT = "1.7.1"
uWSGI = "^2.0.19"
sentry-sdk = "^1.0.0"
numpy = "^1.20.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pep8 = "^1.7.1"
flake8 = "^3.8.4"
autopep8 = "^1.5.5"
isort = "^5.7.0"
pipdeptree = "^2.0.0"
mypy = "^0.812"
django-stubs = "^1.7.0"
djangorestframework-stubs = "^1.4.0"
flake8-isort = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py36', 'py37']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
include_trailing_comma = true
line_length = 88
multi_line_output=3
